cppo: only compare `OCAML_VERSION` as tuples
To test if the version is greater or equal than 4.03.0,
The code base uses variants of `OCAML_MAJOR >= 4 && OCAML_MINOR >= 3`.
This is incorrect after 5.0.0. The right way is to use lexicographic
comparison, in extension:

    (OCAML_MAJOR > 4) || (OCAML_MAJOR == 4 && OCAML_MINOR >= 3)

But fortunately `cppo` supports tuples natively (thanks @Julow for the
tip!), so we can express that as just `OCAML_VERSION >= (4, 3, 0)`.

In addition, this commit standardizes the order in which comparisons are
done:

- conditionals with an else are written with >=:

    #if OCAML_VERSION >= xxx
      ...
    #else
      ...
    #endif

- polyfills for former versions of ocaml are written with <:

    #if OCAML_VERSION < xxx
      ...function introduced in xxx...
    #endif