Pull source as git submodules by default
`duniverse pull` will now clone the source code and add each of
the vendored packages as a git submodule. It does this by adding
the submodule commit objects after the source pulls have been done.
This is a lot faster than the default `git submodule add` which
requires setting up remote tracking branches for each of the
remotes.

This adds a `--no-submodules` option to `pull` to just do a
source code clone.