Add support for a tools config stanza and subcommand
This records the versions of tools required by the repository
in dune-get, with a view to making them available from a new
`duniverse tools` subcommand.

The `tools` subcommand sets up an environment in _ocaml/bin
which has every binary that the project needs to build. This
includes dune and OCaml, as well as mdx, merlin, odoc etc.

The feature isn't quite complete yet, as we still need to flesh
out the support for what happens when opam is available, and
also to add a `duniverse exec` command to set the PATH
appropriately.  These will follow in a later PR.