Fix tag extraction from old janestreet archive urls
Now that the classify package function is called in more places
we need it to support these.

It might actually be a better idea to use a different, more specialized
function here instead since we don't care about the tag when filtering
packages. It's just convenient now to share that logic.