Use create_process instead of execvp
On Windows, execvp creates a separate process, which can confuse dune
when it is waiting for the mdx process. This causes it to miss that the
output file has been created. This affects mdx's test suite and
possibly users of the mdx stanza. Instead we create a process and
immediately wait for it in the parent.

See #295, ocaml/dune#3849.