Remove unecessary subcommands update and opam-install
Update shouldn't be required anymore in its current form. We might
re-implement a similar command if one wants to re-run the lock step but
not for all deps. That's not what the command does atm and it will make
even less sense once we start pinning to tarballs since it only applied
to packages resolved to git repos.

Opam-install isn't required anymore because the lock step will fail if a
package doesn't build with dune, meaning the "opamverse" is always
empty.