Do not use patch_env anymore
`Env.without_cmis` has an issue in 4.04 and 4.05: it corrupts the
environment, making the toplevel unusable after (see ocaml/ocaml#1223).

The `patch_env` function in mdx makes it a noop instead.
This prevents corruption, but will display different results in some cases.

This commit removes that workaround. This has two benefits:

- this removes a complex piece of code, which relies on unsafe behavior
- this aligns the mdx behavior for recent compilers (after 2017) with
  what the toplevel does in these corner cases