Discard compiler variants unless explicitly required
This was causing the solver to select unreleased ocaml versions
sometimes because their dependency could be met using packages such as
`ocaml-variants.4.11.3+trunk`.

The variant selected this way made no particular sense as well since
the solver was picking the highest version number according to the
`+<variant-feature>` suffixes.

The solution is now required to include an `ocaml-base-compiler` unless
one the local packages explicitly depend on an `ocaml-variants` package.
In the future, we'll have to extend this to catch dependencies to
`ocaml-option-*` packages in the future!

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>