Use version field of local package
If a root package has a "version" field in its opam file, use that
version instead of "zdev". This allows breaking some circular
test-dependencies.

Example use case: suppose we are building a monorepo with core and its
test-dependencies. core.v0.14.1 uses async in its tests (this is not
expressed in the public opam metadata), but async depends on core.
Adding `"async" {with-test}` fails in the latest release of
opam-monorepo because async requires core but no async version is
compatible with `core.zdev`. On the other hand, if opam-monorepo knows
that the root core package is at version v0.14.1, it can (correctly)
select async.v0.14.0.