Fix dune-project lang version detection and update to be compatible with windows
This used to crash under some circumstances on windows machines because
it wouldn't handle CRLF properly.
The new code uses regexp to properly extract the version without being
sensitive to leftover CR and replace just the version number, without
modifying the rest of the file, including of course, CRLF.

I also updated the code that appends a lang stanza even though it's
unlikely we'll need it again for a very long time. It now tries to
respect the newline convention found inside the file, default to
good old `\n` if it can't figure it out.

I added a bunch of unit tests for this in the process.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>