Recognise the metadata field on Odoc code blocks
The language tag is used as the header and the other field is for
labels.

If the metadata is absent, the code block is run as OCaml. This is
backward compatible.

Until now, Mdx would strip it from the original file by ignoring it at
parsing and not outputting it back.

This is a rebased version of #339 upon which improvements applied.

Resolves #339