Fix restoring stderr correctly
When running `ocaml-mdx` it would output warnings to the "corrected"
files when running with `--output -`. Since warnings are emitted on
stderr this should not be happening, which means that something was
messing with stderr. It turns out yes, there is code which modifies
stderr but also takes care of restoring it. Unfortunately a typo caused
the backup to be stderr instead so restoring would redirect stderr to
stdout.