Do not trigger warning if there is whitespace after `;;`
OCaml toplevel ignores trailing whitespace so it would make sense to
just ignore it in MDX as well.

Unfortunately Astring does not have `rtrim` so both prefix and suffix
are removed, but except for additional useless computation load it
doesn't make a difference since we only check for the suffix.