Move lockfiles extracting function to Repo
With the fix to the lockfile path inference, we now allow users
to have several lockfiles in their repositories.

This function will likely prove more and more useful so I think it makes
sense to have it here.

As a follow up I'll look into eventually making the depext subcommand
install depexts for all local lockfiles.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>