Improve cram test
This moves the dune file up so it will apply to all future tests
as long as turning the test into a cram folder, incuding the files
to the tree rather than generating them as part of the test.

`cd`ing into the folder to try running commands is now possible which
should help debugging issues.
We still `cat` the content of the files so that the `run.t` file is self
contained and one does not have to browse files to fully understand the
test.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>