Use `Stdlib.Result` implementations
Since our minimum version is 4.08 and our module just adds labels to the
functions (thus functioning like `ResultLabels` if that existed) it
might be somewhat more elegant to just to defer to the built-in
implementations.

This makes it a bit clearer that we're not doing anything different from
the familiar `map`/`bind`/`map_error` implementations.