Use the specified compiler constraint to override chosen constraints
The compiler constraint should be applied after the other constraints
but, it is possible that the set of constraints already contains the
compiler (because a previous run already determined the compiler).

In such case, the compiler constraint should be added successfully but
only if it exactly matches what the previous constraint already
determined.

This fixes an issue where specifying the compiler on the command line
would lead to the constraint being applied twice and failing on the
second time because it was already in place.

It is being cautious to make sure not to overwrite the previously
picked ocaml compiler with a different constraint, so if there is
something unexpected going on it will still throw an error instead of
silencing it and continuing.