Add a Serial_shape module to help with de/serialization
For now handles opam values and Cmdliner arg converters.
The Cmdliner part is a bit harder to deal with due to the
lack of structure of the cmdliner arguments which are just parsed
from raw strings. We might have to later enforce a more structured
format to work around the ambiguities in the cmdliner parsers.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>