Patch and redirect available directives
They have to be redirected because by default they write to stdout, but
we capture stderr (and in 4.14 they write to stderr). Unfortunately,
just overwriting them in the directives Hashtbl does not work, since
the ordering of additions is somewhat undefined so they might be
overwritten. A safer way is to create new directives and rewrite those
to be evaluated to them.

Also, `"use"` is a builtin directive, not from findlib so needs to be
excluded to get the error message printed.