Introduce new syntax for explicit block type
This introduce a new, more explicit labels `block-type=...` which
is meant to replace the four separate labels `cram`, `ocaml`, `toplevel`
and `include` in the future.

It is added for a transition period at the end of which the four labels
of the apocalypse should be removed in favor of this new one.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>