Allow empty lines in front of Cram commands
This adds back the `vpad` functionality which adds newlines in front of
cram blocks.

It also removes the special-casing of not printing the indentation in
some cases, according to the cram spec[0] the output has to start with two
spaces (the padding) and then the output (which is empty in that case).

[0]: https://bitheap.org/cram/