mdx_test: Lazily initialize OCaml toplevel
This changes the OCaml toplevel to be a lazy value, which is only
forced when needed to run tests in an OCaml block.

The motivation behind this change is to avoid stderr output from the
init process (e.g., errors while registering pretty-printers) spilling
into the test executable's output, which can trip up some build tools.