Avoid infinite loop in lexer on unclosed code block
An unclosed code block will cause the lexer to loop infinitely in the
'block' rule, because 'eof' is part of the 'eol' rule and it'll keep
matching the last rule.

Add a separate rule for eof that terminates immediately,
the way the first rule was written it would expect closing quotes after
eof which would never match eof.

Signed-off-by: Edwin Török <edwin@etorok.net>